Add configuration tools

+++ contracts/DAOInterface.sol
@@ -24,42 +24,40 @@ interface DAOInterface {
         address creator;
     }
 
-
-
-
     /// @dev This function is used to send ether back
     /// to the DAO, it can also be used to receive payments that should not be
     /// counted as rewards (donations, grants, etc.)
     /// @return Whether the DAO received the ether successfully
-    function receiveEther() external returns(bool);
+    function receiveEther() external returns (bool);
 
     function newProposal(
         address _recipient,
-        uint _amount,
-        uint _debatingPeriod
-    ) external returns (uint _proposalID);
+        uint256 _amount,
+        uint256 _debatingPeriod
+    ) external returns (uint256 _proposalID);
 
     /// @notice Vote on proposal `_proposalID` with `_supportsProposal`
     /// @param _proposalID The proposal ID
     /// @param _supportsProposal Yes/No - support of the proposal
     /// @return _voteID vote ID.
-    function vote(
-        uint _proposalID,
-        bool _supportsProposal
-    ) external returns (uint _voteID);
+    function vote(uint256 _proposalID, bool _supportsProposal)
+        external
+        returns (uint256 _voteID);
 
-    function executeProposal(
-        uint _proposalID
-    ) external returns (bool _success);
+    function executeProposal(uint256 _proposalID)
+        external
+        returns (bool _success);
 
     /// @notice Get my portion of the reward that was sent to `rewardAccount`
     /// @return _success the call was successful
-    function getMyReward() external returns(bool _success);
+    function getMyReward() external returns (bool _success);
 
     /// @notice Withdraw `_account`'s portion of the reward from `rewardAccount`
     /// to `_account`'s balance
     /// @return _success the call was successful
-    function withdrawRewardFor(address _account) external returns (bool _success);
+    function withdrawRewardFor(address _account)
+        external
+        returns (bool _success);
 
     event ProposalAdded(
         uint256 indexed proposalID,
@@ -67,5 +65,9 @@ interface DAOInterface {
         uint256 minQuorum,
         uint256 votingDeadline
     );
-    event Voted(uint indexed proposalID, bool position, address indexed voter);
+    event Voted(
+        uint256 indexed proposalID,
+        bool position,
+        address indexed voter
+    );
 }
